src/login/requirements.txt
fastapi
sqlalchemy
pyjwt
psycopg2
pydantic
uvicorn
